+++ src/dataflow/concatenate_h5mu/script.py
@@ -371,6 +371,17 @@ def main() -> None:
         raise ValueError("--mode 'move' requires --input_ids.")
 
     n_processes = meta["cpus"] if meta["cpus"] else 1
+
+    if par["modality"]:
+        par["modality"] = set(par["modality"])
+        if not par["modality"].issubset(mods):
+            mods_joined, input_mods_joined = ", ".join(mods), ", ".join(par["modality"])
+            input_mods_joined = ""
+            raise ValueError(
+                f"One of the modalities provided ({input_mods_joined}) is not available in the input data {mods_joined}"
+            )
+        mods = par["modality"]
+
     concatenate_modalities(
         n_processes,
         list(mods),