+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## NEW FUNCTIONALITY
 
+* `dataflow/concatenate_h5mu`: add `modality` parameter (PR #).
+
 * `filter_with_scrublet`: add `expected_doublet_rate`, `stdev_doublet_rate`, `n_neighbors` and `sim_doublet_ratio` arguments (PR #974).
 
 * `feature_annotation/aling_query_reference`: : Added a component to align a query and reference dataset (PR #948, #958).